Path refactoring for the internal packages
