+++ internal/stats/client.go
