+++ internal/stats/client_test.go
