+++ internal/storage/badger/store.go
@@ -16,8 +16,8 @@ import (
 
 	"github.com/dgraph-io/badger/v2"
 	badger_pb "github.com/dgraph-io/badger/v2/pb"
-	"github.com/flipkart-incubator/dkv/internal/server/stats"
-	"github.com/flipkart-incubator/dkv/internal/server/storage"
+	"github.com/flipkart-incubator/dkv/internal/stats"
+	"github.com/flipkart-incubator/dkv/internal/storage"
 	"github.com/flipkart-incubator/dkv/pkg/serverpb"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zapcore"
